Refactor: - Move accelerometer characteristic into seperate file.
